requirements.txt
# Core packages (from HVACRange_Deviation_Forecast/venv, Python 3.12.10)
# v3-0 追加パッケージ:
#   faiss-cpu==1.13.2   : FAISS インデックス（テキスト埋め込みの近傍探索）
#   openai              : オプション（OpenAI API バックエンドを使う場合）
faiss-cpu==1.13.2
# PyTorch (CUDA 12.4) は別途インストール:
#   pip install torch==2.6.0+cu124 torchaudio==2.6.0+cu124 torchvision==0.21.0+cu124 --index-url https://download.pytorch.org/whl/cu124

accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.5.0
debugpy==1.8.20
decorator==5.2.1
Deprecated==1.3.1
dill==0.4.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
granite-tsfm @ git+https://github.com/ibm-granite/granite-tsfm.git@8c9ac38cf3114d909e59e29c73ae4c0daf9d6181
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
imbalanced-learn==0.14.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
lightgbm==4.6.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
narwhals==2.16.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.63.1
numpy==1.26.4
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
peft==0.18.1
pillow==12.1.1
platformdirs==4.5.1
plotly==6.5.2
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.0
pycparser==3.0
Pygments==2.19.2
pynndescent==0.6.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.1.15
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.3.2
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scikit-metrics==0.1.0
scipy==1.17.0
seaborn==0.13.2
Send2Trash==2.1.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.5
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==4.56.0
typer==0.23.0
typer-slim==0.23.0
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.11
uri-template==1.3.0
urllib3==2.6.3
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
